requirements.txt
Appium-Python-Client==0.31
pytest==4.0.1
six==1.11.0